Adding type GET to minifiedAjax
Hello,

i'm adding type GET to minifiedAjax,
because when customizing script.blade.php and using thrid party tools like metronic datatable helper,
the config could be merged with a default one in the process. 

The default type was POST in my case so it's not working.
I can always set this in my config, but minifiedAjax is for GET only, so i it's relevant to hard code the type of the request

Thanks.