Some Updates
- queryString:
We have to be able to see the id of the dataTable by looking at the ajax request, since we might want to handle more than one data tables in one route.

- Custom JS functions:
The code now allows jQuery and JS codes to be printed as literal functions.

- SmartDataTables:
Also we want to specify if the request will be handled from the new SmartDataTables service.

So,
Let me explain what SmartDataTables mean for me. My company uses a s* ton of datatables and I am so tired of copy pasing a new route/method over and over again.

SmartDataTables simply adds a field called `$isSmartDataTable` to the `abstract DataTable` class which is by default `false`. On your child class you can overwrite this.

Ex:
```php
    /**
     * Dynamically detects which DataTable to initialize for backend processing and responds.
     *
     * @var boolean
     */
    protected $isSmartDataTable = true;
```

By doing so you let the dataTable add a `smartDataTable` field to the ajax request which has the value of the class which is a "SmartDataTable". If we were to make a UsersDataTable and mark it as "smart". The query will have a parameter like this: `smartDataTable=App\DataTables\UsersDataTable`.

The next step is to check all web requests via a Middleware if the contain the parameter `smartDataTable`. If they do, we will simply resolve the data table class and respond to it accordingly. (Just like the `render` method does.)

```php
        if (!$request->has('smartDataTable')) {
            return $next($request);
        }

        $dataTable = app($request->smartDataTable);

        if ($request->ajax() && $request->wantsJson()) {
            return app()->call([$dataTable, 'ajax']);
        }
```